reuse more code
